requirements.txt
fastapi
uvicorn
python-dotenv
motor
dnspython
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
python-multipart

httpx
google-generativeai
